Add access to admin for all ticket crud

+++ TaskManagerMVP/Controllers/TicketsController.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -23,17 +24,29 @@ namespace TaskManagerMVP.Controllers
             _context = context;
         }
 
+        protected async Task<string> GetCurrentUserId()
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return userId;
+        }
         // GET: Tickets
         public async Task<IActionResult> Index(string ticketProject)
         {
+            var userId = await GetCurrentUserId();
             //Use LINQ to get list of projects
             IQueryable<string> projectsQuery = from t in _context.Projects
                                                orderby t.Name
                                                select t.Name;
-            //allow for search options
+            //Set query to get tickets only for logged in user
             var tickets =  _context.Tickets
-                              .Include(t => t.Project).Include(t => t.TicketPriority).Include(t => t.TicketStatus).Include(t => t.TicketType).Include(t => t.User); ;
-            if (!String.IsNullOrEmpty(ticketProject))
+                              .Where(x => x.UserId == userId)
+                              .Include(t => t.Project).Include(t => t.TicketPriority).Include(t => t.TicketStatus).Include(t => t.TicketType).Include(t => t.User); 
+            if (User.IsInRole("Admin")) //query to get ALL tickets if admin
+            {
+                tickets = _context.Tickets
+                             .Include(t => t.Project).Include(t => t.TicketPriority).Include(t => t.TicketStatus).Include(t => t.TicketType).Include(t => t.User);
+            }
+            else if (!String.IsNullOrEmpty(ticketProject))
             {
                 tickets = _context.Tickets
                     .Where(s => s.Project.Name!.Contains(ticketProject)) 
@@ -53,6 +66,7 @@ namespace TaskManagerMVP.Controllers
         // GET: Tickets/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var userId = await GetCurrentUserId();
             if (id == null)
             {
                 return NotFound();
@@ -69,18 +83,28 @@ namespace TaskManagerMVP.Controllers
             {
                 return NotFound();
             }
-
-            return View(ticket);
+            //I don't know how to check BEFORE I run the query - but this doesn't seem efficient
+            if (ticket.UserId == userId || User.IsInRole("Admin"))
+            {
+                return View(ticket);
+            }
+            else
+            {
+                return Forbid(); //only admins and users should see their own tickets
+            }
         }
 
         // GET: Tickets/Create
         public IActionResult Create()
         {
+            var userId = GetCurrentUserId();
+
             ViewData["ProjectId"] = new SelectList(_context.Projects, "Id", "Name");
             ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name");
             ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name");
             ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name");
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "UserName");
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName");
+            
             return View();
         }
 
@@ -91,6 +115,7 @@ namespace TaskManagerMVP.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Description,StartDate,EndDate,UserId,ProjectId,TicketTypeId,TicketStatusId,TicketPriorityId,IsActive")] Ticket ticket)
         {
+            var userId = GetCurrentUserId();
             if (ModelState.IsValid)
             {
                 _context.Add(ticket);
@@ -101,13 +126,23 @@ namespace TaskManagerMVP.Controllers
             ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name", ticket.TicketPriorityId);
             ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name", ticket.TicketStatusId);
             ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name", ticket.TicketTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "UserName", ticket.UserId);
+           
+            if (User.IsInRole("Admin"))
+            {
+                ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", ticket.UserId);
+
+            }
+            else
+            {
+                ViewData["UserId"] = userId;
+            }
             return View(ticket);
         }
 
         // GET: Tickets/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            var userId = await GetCurrentUserId();
             if (id == null)
             {
                 return NotFound();
@@ -118,12 +153,22 @@ namespace TaskManagerMVP.Controllers
             {
                 return NotFound();
             }
-            ViewData["ProjectId"] = new SelectList(_context.Projects, "Id", "Name", ticket.ProjectId);
-            ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name", ticket.TicketPriorityId);
-            ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name", ticket.TicketStatusId);
-            ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name", ticket.TicketTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "UserName", ticket.UserId);
-            return View(ticket);
+            //I don't know how to check BEFORE I run the query - but this doesn't seem efficient
+            if (ticket.UserId == userId || User.IsInRole("Admin"))
+            {
+                ViewData["ProjectId"] = new SelectList(_context.Projects, "Id", "Name", ticket.ProjectId);
+                ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name", ticket.TicketPriorityId);
+                ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name", ticket.TicketStatusId);
+                ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name", ticket.TicketTypeId);
+                ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", ticket.UserId);
+                return View(ticket);
+            }
+            else
+            {
+                return Forbid(); //only admins and users should see their own tickets
+            }
+            
+            
         }
 
         // POST: Tickets/Edit/5
@@ -133,47 +178,58 @@ namespace TaskManagerMVP.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Description,StartDate,EndDate,UserId,ProjectId,TicketTypeId,TicketStatusId,TicketPriorityId,IsActive")] Ticket ticket)
         {
+            var userId = await GetCurrentUserId();
             if (id != ticket.Id)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            //if userId doesn't match, or if not admin then don't allow for edit
+            if (ticket.UserId == userId || User.IsInRole("Admin"))
             {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(ticket);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!TicketExists(ticket.Id))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(ticket);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!TicketExists(ticket.Id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["ProjectId"] = new SelectList(_context.Projects, "Id", "Name", ticket.ProjectId);
+                ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name", ticket.TicketPriorityId);
+                ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name", ticket.TicketStatusId);
+                ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name", ticket.TicketTypeId);
+                ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", ticket.UserId);
+                return View(ticket);
             }
-            ViewData["ProjectId"] = new SelectList(_context.Projects, "Id", "Name", ticket.ProjectId);
-            ViewData["TicketPriorityId"] = new SelectList(_context.Priorities, "Id", "Name", ticket.TicketPriorityId);
-            ViewData["TicketStatusId"] = new SelectList(_context.Statuses, "Id", "Name", ticket.TicketStatusId);
-            ViewData["TicketTypeId"] = new SelectList(_context.TicketTypes, "Id", "Name", ticket.TicketTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "UserName", ticket.UserId);
-            return View(ticket);
+            else
+            {
+                return Forbid();
+            }
+            
         }
 
         // GET: Tickets/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            var userId = await GetCurrentUserId();
             if (id == null)
             {
                 return NotFound();
             }
 
+            
             var ticket = await _context.Tickets
                 .Include(t => t.Project)
                 .Include(t => t.TicketPriority)
@@ -185,8 +241,17 @@ namespace TaskManagerMVP.Controllers
             {
                 return NotFound();
             }
+            //if userId doesn't match, or if not admin then don't allow for delete
+            if (ticket.UserId == userId || User.IsInRole("Admin"))
+            {
+                return View(ticket);
+            }
+            else
+            {
+                return Forbid();
+            }
 
-            return View(ticket);
+            
         }
 
         // POST: Tickets/Delete/5
@@ -194,10 +259,16 @@ namespace TaskManagerMVP.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var userId = await GetCurrentUserId();
+
             var ticket = await _context.Tickets.FindAsync(id);
-            _context.Tickets.Remove(ticket);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (ticket.UserId == userId || User.IsInRole("Admin"))
+            {
+                _context.Tickets.Remove(ticket);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return Forbid();
         }
 
         private bool TicketExists(int id)
